renaming vue components
